feat: 통계테이블 컴포넌트 스토리북 추가 (#22)
